+++ classes/table/all_notices.php
@@ -319,7 +319,7 @@ class all_notices extends table_sql implements renderable {
         return html_writer::link(
             "#",
             get_string('view'),
-            ['class' => 'notice-preview', 'data-noticecontent' => $awareness->get('content')]
+            ['class' => 'notice-preview', 'data-noticecontent' => helper::render_content($awareness)]
         );
     }
 }